+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Next
+
+- Upgraded tooling.
+- Replaced `jsonld-streaming-parser` and `jsonld-streaming-serializer` dependencies with `jsonld`.
+
 ## [v0.4.1](https://github.com/NoelDeMartin/soukai-solid/releases/tag/v0.4.1) - 2021-01-25
 
 ### Changed